yo: Add SSH console options
The SSH console used by OCI needs some special settings. I've had these
settings on my development machine's ~/.ssh/config for some time, but
since installing it on my personal machine, I've seen that the config
needs to be set by Yo in order for users to have an easy time of it.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>