ssh: Improve compatibility of SSH console options
From user feedback, it seems that using a space in the SSH option syntax
is not as well supported as using an equals sign (which is how all the
other options are specified, anyway). So make sure we use an equals
sign.

Further, revert to using the older name "PubkeyAcceptedKeyTypes" rather
than "PubkeyAcceptedAlgorithms" -- while the latter is the newer, more
"correct" name, it was renamed in OpenSSH 8.5, and the former remains as
an alias. For maximal compatibility, we use the older alias.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>