api: Don't bail out on multiple VNIC attachments
I used the one() function since I have not encountered instances with
multiple VNIC attachments, and I didn't know how the behavior would
impact the use case of SSH. However, I've had a user report that simply
using the first attachment would have worked fine for them, and Yo
instead crashes, which is not ideal. Rather than fail on this case,
let's just warn on it and use the first VNIC.

I've added a noisy warning so that users know to let me know if they
encounter this issue. The warning won't be so bad for people who have
multiple VNICs with no trouble: since the IP is cached, the warning
isn't printed frequently.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>