yo console: allow more instance states
Running yo console should be permitted in almost any state, except if
the instance is actually stopped or terminated. Of course, the STARTING
and PROVISIONING states are frequently a stale cache, but even STOPPING
or TERMINATING should be allowed, as it may be necessary to look at the
console while an instance stops.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>