Use entry points for extensions
Yo's not-so-advertised extension API makes it possible to customize
certain elements of the program. It's undocumented because there's no
actual API, but if you know the code, it's possible to hook into things
and modify their behavior. This is used in Oracle to customize some
links to documentation and version messages, as well as provide some
internal task scripts.

However, it was required to module names into a configuration element.
Python comes with entry_points specifications in packages, so this
commit adds support for that, in addition to the list of module names.
Ideally, we'll get rid of the old way, but for now it's fine to have
both.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>