Fix TypeError on startup for Python 3.8 & 3.9
I hoped to use the most recent API for entry points, and thus preferred
importlib.metadata unless it is unavailable. Unfortunately, I missed the
small italic message at the top of the documentation:

> Changed in version 3.10: importlib.metadata is no longer provisional.

Unfortunately, entry_points() doesn't take a group parameter in the
provisional API, and it returns a completely different type. So running
my well-intentioned code yields the following on 3.8 and 3.9:

  Traceback (most recent call last):
    File "/opt/fajita/bin/yo", line 8, in <module>
      sys.exit(main())
    File "/opt/fajita/venv/lib64/python3.9/site-packages/yo/main.py", line 3162, in main
      _extend(ctx)
    File "/opt/fajita/venv/lib64/python3.9/site-packages/yo/main.py", line 3148, in _extend
      for entry_point in entry_points(group="yo.extensions.v1"):
  TypeError: entry_points() got an unexpected keyword argument 'group'

I suppose this is programmer error: a failure to closely read the docs,
along with a failure to include this particular piece of code in my
small tox test suite which runs on each Python from 3.6 through 3.11.
However, I wish that the documentation were clearer about the
provisional API. I've long held the belief that, if it works on the
oldest Python version and the newest, then it will work on all in
between. This is an unfortunate counterexample, and I'll have to
reexamine that belief going forward.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>