Make cache file updates atomic
Writing the cache file concurrently to another reader can have the
unfortunate consequence of the reader getting a partial read, and thus
crashing with a JSON decode error. To avoid this, write the file to a
temporary filename and then use rename(), which will atomically replace
the old file. This means the reader gets no partial result: either the
old or the new value, but never anything else.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>