List instances prior to state changes
For commands like terminate, start, stop, etc, it is very important to
have an up-to-date view of instance states. A long time ago, the OCI SDK
and API were slow enough that avoiding any API call was a good idea, but
now with selective service imports, and performance improvements, the
calculus isn't the same.

Let's go ahead and just call list_instances() before any of these
commands. This will ensure that the instance selection logic gets the
most accurate data, and we won't go off trying to do actions on
instances based on outdated information.

To do this, we modify the get_matching_instances() logic, and it's
enough to update MultiInstanceCommand: every instance action command,
including terminate, eventually inherits from that.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>