yo list: add --ip
The yo ip command is designed to be a fast, cache-only operation where
possible. This means that it never refreshes the cache. Let's add an
option to display the IP address of the instance in yo list too. It's a
bit odd that it's not there to begin with. This ensures that if people
are running yo ip as a replacement for yo list, they have an option that
refreshes the cache by default.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>