Add bulk IP lookup functionality
Despite being a very important feature, getting an IP address associated
with an instance is not straightforward. You must lookup the
"VnicAttachment" list for an instance, and then query each individual
"Vnic" object to get the IP address.

The existing get_instance_ip() function does this for one instance at a
time: query all attachments for a given instance, then query the vnic.
However, we have some commands that are now fetching IPs for all
instances, e.g. "yo list --ip". It's inefficient to do all these API
calls in serial, and if we were to simply list all the attachments at
once (without filtering by instance) we could improve performance more.

So, add a get_all_instance_ips() functionality, which first lists *all*
attachments, and then delegates the "get_vnic()" call to a thread pool
for whatever instance IDs it needs. The function respects and adds to
the cache, so that we never query more than we need, and repeated calls
won't query more.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>