instance naming: make it consistent and provide global config
I've been putting off this change for sooo long. Yo is very opinionated
on the instance name, and I think for the majority of people launching
temporary instances in a shared compartment, this is the correct way to
go.

However, there are plenty of cases where you don't want to do the whole
"username prefix" thing. If you don't share a compartment, there's not a
compelling reason to do it at all. And there are some cases where you
need a specific name.

We've already provided the "--exact-name" argument on a lot of
sub-commands. However, the block-volume management commands did not
provide this argument. This was recently pointed out to me, so I needed
to add the argument there anyway.

This change goes a bit further than that. In addition to having the
--exact-name argument, I've added a short-hand "-E" which does the same
thing. And I've also added a global configuration option which implies
--exact-name all the time. This means that, if you don't like this
particular behavior, you can set that in your config file and forget
about it.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>