Fix created_by of volumes and images
Some tenancies do not have the Oracle-Tags.CreatedBy field, as
documented elsewhere in Yo. This is how Yo filters resources, but a new
functionality was added for this special case, where Yo can tag its
resources and fall back to this tag.

However, that fallback was only implemented for instances. Other
resources, like images and block volumes, have the same problem and
weren't fixed. This leads to frustrating situations where a block volume
can be created, but then Yo cannot see it or attach it to anything.

So, update the logic to fall back to the special Yo tag on all these
resource types. Invalidate caches to ensure they are updated.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>