Skip fchmod on cache if unavailable
This was always a nice-to-have hardening step similar to the required
permissions on SSH key files, but it appears Windows permissions won't
work with this.

Thankfully, the cache is not nearly as sensitive as SSH keys... it was
just nice to keep it out of prying eyes. So let's drop the fchmod when
it's not available.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>