Do not return after invalidating cache
The load_cache() function is designed to work even if there is no cache
file, and the cache is just an empty dict. The return statement here is
unnecessary, so long as we reset the cache to the empty dict value.

The subsequent logic sets the "last_checked_for_update" attribute, and
if we return early, we don't get that attribute set, and thus "yo list"
could crash. This can occur on an upgrade or on a changed
"resource_filtering" config value, if the next command is "yo list".

So, simply reset the cache variable.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>