Improving wait and "yo launch" experience
We're using the "rich" library, which is quite powerful and pretty, but
we're unfortunately under-utilizing it, and the print-outs that Yo makes
are scattered and not standardized. This means that the experience for
all of this is a bit cluttered: messages getting printed out here,
there, everywhere.

What's more, some useful information is *almost* provided, but not. For
instance, each logged message could provide some timestamp info. And the
"wait for X" messages could tell you how much time elapsed, but they
don't currently.

So for this change, we update all the progress displays to be more
sensible. There's no need for a progress bar (since we can't actually
measure progress toward the final goal), so switch it to a spinner.
Include the time elapsed and remaining (both).

Further, use console.log() which includes a timestamp for each
message, and a neat filename:line number display at the right-hand side.
This provides a bit of structure to the otherwise unstructured output,
and gives the missing timing information. The overall experience feels
fresh and more informative, while simultaneously being less cluttered.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>