Replace MultiInstanceCommand progress bar
All the rest of the progress bars are now spinners. This is nicer
looking since we don't really have a good way to predict how long
something will take, and so it looked a bit silly. While the
MultiInstanceCommand can indicate its progress since it's processing a
known list of items, it would be better to show this progress without
the bar, so everything is consistent.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>