api: Add alt_name field for YoVolume
Boot volumes get automatically named "instance name (Boot Volume)",
which is not a good name for the CLI. I assumed that "boot volume" was
added by Yo for the table layout at first, so I tried to use the
instance name on its own. Add an alt_name field to contain just the
instance name so we can support lookup based on that as well.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>