ssh: provide 60-second grace period for timeout warning
The timeout warning can be super helpful, but 90% of the time it appears
for no reason except that we tried to connect via SSH too soon. It can
worry users unnecessarily. Let's have a 60 second grace period, and not
show the warning until it's at least that long since boot.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>