yo compat: improve compatibility table output
The compatibility table is rarely used, but it's a good diagnostic.
However it's difficult to read because the lines blend together. Use
alternating bold font to help visually distinguish the lines, which
makes it much easier to read.

In addition, increase the width of the shape name column (make it
adjustable too). Add filtering options so that you don't need to have a
giant console all the time, just to look at it. And finally, add a mode
which will print the image names rather than OS names. This is useful
but produces huge output, it's generally best to filter.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>