Use natural sort for image names
By doing so, numbers are treated separately from strings and they are
compared based on their integer value, rather than their alphabetical
characters. Thus, 8.10 is considered "greater than" 8.9, whereas it
would not be in a standard alphanumeric sort.

This sort is to be used for the image listing, as well as the actual
routine which selects an image for launch.

I prefer this rather than the much simpler approach of sorting by the
"time_created" of the image, because in my experience, images are named
based on the date and version info when they were created. It's totally
possible for an old image to be uploaded after a newer one. Sorting
based on the image's timestamp would make it difficult to resolve that
situation, whereas the natural sort easily selects the latest based on
the embedded dates in the name.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>