Add contrib/yo.zsh
This adds an alternative zsh completion definition for yo that includes
tags, descriptions, and argument groups, beyond the bash-style argument
listings and explanations offered by the python argcomplete provided
completions.

Because they are not the default, for now the user would need to install
them as any other supplementary completion definition by including this
file in their $fpath with the filename "_yo" and re-running compinit.

jq is used to parse the yo cache files when it is installed. Otherwise,
argument listings for some words (isntances, shapes, etc.) are not
available.

Signed-off-by: Ronan Pigott <ronan.pigott@oracle.com>