docs: quick fix generation script and update
The generation script updates the command documentation and especially
the index, but it always added an extra two newlines which I removed
before committing. Get rid of those spare two spaces in the generator
script, and re-run it to pick up a recent change in the "yo cache-clean"
command description.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>