Add "yo script" command
This is a tentative new feature which would allow users to (hopefully
quite easily) automate certain instance actions which are too
complicated to make a bash script for (since Yo is notoriously bad for
that). The script is executed with the YoCtx object in a global variable
named "ctx". A sample hello world script would be:

    #!/usr/bin/env -S yo script
    from yo.api import YoCtx

    def main(ctx: YoCtx):
        ctx.con.print("Hello world")

    if __name__ == "__main__":
        main(ctx)  # noqa

Note that the shebang line would allow the user to directly run the
script, although it is also perfectly acceptable to run it as "yo script
foo.py".

Honestly, the state of Yo is that a lot of the functionality is buried
into the sub-classes of YoCommand, so it may not actually be very easy
to write indpendent scripts for Yo. In the future I may make small
changes that move functionality into the YoCtx so that it works better
as an API.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>