yo launch: move image selection into YoCtx
One of the necessary pieces of the puzzle for an API that makes
it (relatively) easy to launch instances in a Yo script would be the
ability to select an image by OS, the way that Yo users can. This
functionality is buried in a method of the command implementing
"launch". Let's factor that out into its own YoCtx method that can
easily be used in scripts.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>