yo launch: move most logic into YoCtx
The LaunchCmd contains most of the logic that makes yo "convenient".
That is, selecting images, volumes, shapes, subnets, etc. If you're
going to use Yo as an API, you want to have the convenient launch
commands. Otherwise you might as well just write a script that directly
calls the OCI SDK.

So, this refactor moves basically all of the logic which determines the
"launch config" into YoCtx.get_launch_config(). And an easy wrapper for
getting a launch config and using it is applied. This way an API user
can simply run:

    ctx.simple_launch("my_profile"))

And this would be equivalent to "yo launch -p my_profile". Of course,
there are other things that "yo launch" does after the instance is
created: wait for the instance to launch, SSH, run tasks, etc. This is
outside of the scope of the simple launch API, so that logic stays in
LaunchCmd.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>