task: add the ability to create tasks at runtime
For scripting use cases, it is nice to be able to put a script into a
string literal, customize it a bit, and then create a task based on
that. Yo normally loads tasks by name out of a list of task directories.
Add the ability to create a task with a name and a script. All the same
conveniences of tasks (e.g. dependencies) are available.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>