Better format for OCI ServiceErrors
When service errors are uncaught, they tend to have a lot of information
buried inside a dictionary that's difficult to read. Users don't
necessarily have the familiarity or desire to parse through that and
find the actual message that OCI has for them.

Some of these ServiceErrors are mistakes that the user made (e.g.
trying to attach a disk to an instance, when it is already non-shareably
attached to a different instance). But others are mistakes by the
developer (hello!). Let's extract that information and emphasize it,
while de-emphasizing the traceback as much as possible. By doing this,
users get a chance to read the salient parts of the error and they have
a better chance of handling the errors of the first kind. By
de-emphasizing (but not eliminating) tracebacks, errors of the second
kind can still be handled via bug reports.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>