Import subc library
Subc is a small library I wrote a while back to facilitate simple
argparse based CLI applications which had sub-commands. It is used in
multiple of my personal projects, but Yo is clearly the largest user.

Due to its small size, it's somewhat silly to include as a dependency,
especially if RPM packaging is to be considered. Since I'm the only
author (it has never received a contribution), I fully own the copyright
and can include it here with my developer sign-off. In this first
commit, I include it unmodified from the v0.8.0 release. Subsequent
commits will address type annotations, file headers, static checks, etc.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>