Refactor: split out ssh and task management code
Yo's code has been a bit of a mess in terms of organization, with the
main.py weighing many thousands of lines of code. As tasks become more
complex, they are meriting their own implementation file. And since
tasks depend on the SSH functionality (which also lives in main.py),
this would require a circular dependency. So split out the SSH code into
a separate module, and the task code too.

No functional change is intended.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>