tasks: allow specifying task "-" to clear profile
It's nice to have tasks in your instance profile for the most common
case, but overriding the tasks is not fully possible: you can only add
tasks, not remove them.

Add a blunt instrument for this, by specifying "-" you can instruct Yo
to skip all the tasks specified in the instance profile.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>