Strip whitespace from instance_compartment_id
A user got a quite uninformative 404 error in the _list_ads() call.
Turns out that "instance_compartment_id" had a newline in it, because
they saw the format of "image_compartment_ids" and figured they could
put the instance_compartment_id indented on a newline in the same way.

That's a totally reasonable assumption, and it's just a quirk of how Yo
treats these different parameters that it didn't work. Fix this so
nobody else has to deal with it.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>