yo compat: improve table formatting and fix layout bug
The table format for yo compat left something to be desired. Implement
the color coding and the shading as background colors, which are always
visible. Add visual distinction for the rows with underlines.

This improvement helped me read the table and understand that we had a
bug when "--image-names" was not specified: the images should have been
sorted by os version so that the proper highlights were present. I was
confused why there were aarch64 images present under the GPU OS names.
This bug was the cause. Fix it here as well.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>