Specify custom images with compartment as a prefix
This commit requires the "os" and "image" specifiers to contain the
compartment name as a prefix when specifying custom images. This is
because two users recently encountered cases where a custom image with
OS "Oracle Linux" was added to a compartment they used. In both cases,
the custom image did not have a "OracleTags.Created-By" tag, so Yo
assumed it was official, and selected the custom image rather than the
platform image, which was unexpected.

A prior commit fixed Yo's platform image detection, by expecting
"compartment_id" to be None for platform images. However, it doesn't
really fix the underlying issue, that all images share a single
namespace. Custom images can shadow the names of each other, as well as
the platform images. What's more, custom images cannot currently be
specified by OS, despite that being a useful capability. For instance, I
upload Fedora images to my compartment, and I would like to be able to
specify them by OS/version.

To rectify all this, let's make a rare breaking change. All OS and image
specifiers must now be prefixed by the compartment name, plus a colon,
when they are specifying a custom image. Platform images of course, can
still be specified the same way.

To support this change, we need to update the docs and changelog. We
also need to update the images command to handle the new specifiers, and
the "os" command to output the correct specifiers.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>